courses/pyproject.toml
[tool.poetry]
name = "courses"
version = "0.1.0"
description = ""
authors = ["sky-sky-code <saidyllapashaev@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.3"
aio-pika = "^9.4.0"
redis = ">=4.2.0rc1"
locust = "^2.23.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
